chore(files): Move some components to modals folder, move bonus photon to game.
