+++ src/lib/components/modals/GlobalStats.svelte
