+++ src/lib/components/modals/HardReset.svelte
